Module Wallet.Mnemonic

Mnemonic of 24 common english words from which a key can be derived. The mnemonic follows the BIP-39 spec.

val new_random : Tezos_client_base.Bip39.t

val words_pp : Stdlib.Format.formatter -> string list -> unit

Pretty printer for printing a list of words of a mnemonic.